Move `Config` to `components` module
