+++ audiovlm_demo/__init__.py
@@ -4,6 +4,5 @@ __all__ = [
     "Config",
 ]
 
-from audiovlm_demo.core.components import AudioVLM
-from audiovlm_demo.core.config import Config
+from audiovlm_demo.core.components import AudioVLM, Config
 from audiovlm_demo.ui.panel import AudioVLMPanel
