+++ audiovlm_demo/core/components.py
@@ -2,10 +2,14 @@ import gc
 import io
 import re
 import time
-from typing import Any
+from pathlib import Path
+from typing import Annotated, Any
 
 import librosa
+import tomlkit
 import torch
+from pydantic import AfterValidator
+from pydantic_settings import BaseSettings
 from transformers import (
     AutoModelForCausalLM,
     AutoProcessor,
@@ -14,7 +18,32 @@ from transformers import (
     Qwen2AudioForConditionalGeneration,
 )
 
-from audiovlm_demo.core.config import Config
+from audiovlm_demo.core.utils import resolve_path
+
+_ResolvedPath = Annotated[Path, AfterValidator(resolve_path)]
+
+
+class Config(BaseSettings):
+    """
+    Class to store configuration for the demo.
+
+    Includes paths to downloaded models, among other thnigs.
+    """
+
+    model_path: _ResolvedPath
+    aria_model_path: _ResolvedPath
+    qwen_audio_model_path: _ResolvedPath
+
+    # TODO: I would like to remove the quotes in the return type
+    # annotation without getting a NameError
+    @classmethod
+    def from_file(cls, path: str | Path) -> "Config":
+        path = resolve_path(path)
+        if not path.is_file():
+            raise FileNotFoundError(f"{path} does not exist.")
+
+        with open(path) as file:
+            return cls.model_validate(tomlkit.load(file).unwrap())
 
 
 class AudioVLM:
